+++ ohana-project/src/main/java/fr/isika/cda11/ohana/project/crowdfunding/service/ProjectService.java
@@ -33,7 +33,7 @@ public class ProjectService {
 		
 	}
 	
-	public Project findProjectService(Integer id) {
+	public Project findProjectService(Long id) {
 		
 		return projectRepos.findProjectRepos(id);
 		
@@ -44,10 +44,5 @@ public class ProjectService {
 		return projectRepos.findProjectsRepos();
 		
 	}
-	
-
-
-	
-	
 	
 }
